backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.13",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asyncpg==0.30.0",
    "attrs==25.3.0",
    "certifi==2025.6.15",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "colorama==0.4.6",
    "dnspython==2.7.0",
    "dotenv==0.9.9",
    "email-validator==2.2.0",
    "fastapi==0.115.13",
    "fastapi-cli==0.0.7",
    "frozenlist==1.7.0",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "iniconfig==2.1.0",
    "jinja2==3.1.6",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "multidict==6.6.2",
    "packaging==25.0",
    "pillow==11.2.1",
    "pluggy==1.6.0",
    "propcache==0.3.2",
    "psycopg2-binary==2.9.10",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pygments==2.19.2",
    "pytest==8.4.1",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "requests==2.32.4",
    "rich==14.0.0",
    "rich-toolkit==0.14.7",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.43",
    "starlette==0.46.2",
    "typer==0.16.0",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "upstash-redis>=1.5.0",
    "urllib3==2.5.0",
    "uvicorn==0.34.3",
    "watchfiles==1.1.0",
    "websockets==15.0.1",
    "yarl==1.20.1",
]
